Rename shared_bytes to just shared
