+++ patomic/src/shared.rs
