+++ patomic/src/transaction.rs
@@ -1,0 +1,2 @@
+// Copyright (c) doodspav.
+// SPDX-License-Identifier: MIT OR Apache-2.0
